// UI/ChathireUI/src/app/api/model/job-opening-candidate-profile-map-dto-for-insert.ts
/**
 * REST API
 * Rest Calls
 *
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface JobOpeningCandidateProfileMapDtoForInsert { 
    jobOpeningId?: number;
    candidateProfileId?: number | null;
    appliedDate?: string | null;
    active?: boolean | null;
    candidateProfileMappingStatusId?: number;
    comment?: string | null;
    consultancyUserId?: number | null;
    candidateUserId?: number | null;
    doc?: string | null;
    candidateName?: string | null;
    totalYearsOfExp?: number | null;
    linkedIn?: string | null;
    currentLocationCityid?: number | null;
}